Symmetry adapted pool constructed at c++

+++ src/qforte/find_irrep.h
@@ -1,5 +1,6 @@
 #include <vector>
+#include <cstddef>
 
-int find_irrep(const std::vector<int>& orb_irreps_to_int,
-        const std::vector<int>& spinorb_indices);
+size_t find_irrep(const std::vector<size_t>& orb_irreps_to_int,
+        const std::vector<size_t>& spinorb_indices);
